Correccion entero a decimal de monedas en retenciones RH

+++ Convertidor/Data/Entities/Rh/RH_TMP_RETENCIONES_SSO.cs
@@ -9,14 +9,14 @@
         public string NOMBRES_APELLIDOS { get; set; }
         public string DESCRIPCION_CARGO { get; set; }
         public DateTime FECHA_INGRESO { get; set; }
-        public int MONTO_SSO_TRABAJADOR { get; set; }
-        public int MONTO_RPE_TRABAJADOR { get; set; }
+        public decimal MONTO_SSO_TRABAJADOR { get; set; }
+        public decimal MONTO_RPE_TRABAJADOR { get; set; }
 
-        public int MONTO_SSO_PATRONO { get; set; }
+        public decimal MONTO_SSO_PATRONO { get; set; }
 
-        public int MONTO_RPE_PATRONO { get; set; }
+        public decimal MONTO_RPE_PATRONO { get; set; }
 
-        public int MONTO_TOTAL_RETENCION { get; set; }
+        public decimal MONTO_TOTAL_RETENCION { get; set; }
 
         public string FECHA_NOMINA { get; set; }
